Move workspace related files/structs to new workspace module
