+++ src/common/tests/mod.rs
@@ -1,2 +1,1 @@
 mod sizing_tests;
-mod workspace_tests;
