+++ src/window_manager/tests/test_support.rs
@@ -2,7 +2,7 @@ use super::window_manager::WindowManager;
 use crate::api::MockWindowsApi;
 use crate::configuration_provider::{ConfigurationProvider, Layout};
 use crate::utils::create_temp_directory;
-use crate::workspace_manager::WorkspaceManager;
+use crate::workspace::WorkspaceManager;
 use std::sync::{Arc, Mutex};
 
 impl WindowManager<MockWindowsApi> {
