+++ src/window_manager/tests/window_manager_tests.rs
@@ -1,12 +1,12 @@
 use crate::api::{MockWindowsApi, WindowsApi};
 use crate::common::{
-  Direction, Monitor, MonitorHandle, PersistentWorkspaceId, Point, Rect, Sizing, WindowHandle, WindowPlacement, Workspace,
+  Direction, Monitor, MonitorHandle, PersistentWorkspaceId, Point, Rect, Sizing, WindowHandle, WindowPlacement,
 };
 use crate::configuration_provider::{ConfigurationProvider, Layout};
 use crate::utils::create_temp_directory;
 use crate::window_manager::WindowManager;
 use crate::window_manager::tests::test_support::scrolling_manager;
-use crate::workspace_manager::WorkspaceManager;
+use crate::workspace::{Workspace, WorkspaceManager};
 use std::sync::{Arc, Mutex};
 
 fn vertical_mixed_layout_manager(direction: Direction, target_layout: Layout) -> (WindowManager<MockWindowsApi>, Monitor) {
@@ -645,7 +645,7 @@ fn move_window_to_workspace_when_moving_from_scrolling_to_spatial_removes_strip_
     .unwrap()
     .set_monitor_layout("DISPLAY2", Layout::Spatial);
   manager.reconcile_layouts();
-  let primary_workspace = PersistentWorkspaceId::from(*crate::workspace_manager::tests::primary_active_ws_id());
+  let primary_workspace = PersistentWorkspaceId::from(*crate::workspace::tests::primary_active_ws_id());
   let secondary_workspace = manager
     .workspace_manager
     .active_workspace_ids()
@@ -678,7 +678,7 @@ fn move_window_to_workspace_when_moving_from_spatial_to_scrolling_inserts_strip_
     true,
   );
   MockWindowsApi::place_window(secondary_handle, 2.into());
-  let primary_workspace = *crate::workspace_manager::tests::primary_active_ws_id();
+  let primary_workspace = *crate::workspace::tests::primary_active_ws_id();
   let mut manager = WindowManager {
     configuration_provider,
     placement: Default::default(),
