+++ src/workspace/mod.rs
@@ -1,0 +1,17 @@
+mod transient_workspace_id;
+#[allow(clippy::module_inception)]
+mod workspace;
+mod workspace_action;
+mod workspace_guard;
+mod workspace_manager;
+
+pub(crate) use transient_workspace_id::TransientWorkspaceId;
+pub(crate) use workspace::Workspace;
+pub(crate) use workspace_manager::WorkspaceManager;
+
+#[cfg(test)]
+pub(crate) mod tests {
+  pub(crate) use super::workspace_manager::tests::*;
+
+  mod workspace_tests;
+}
