+++ src/workspace/workspace_manager.rs
@@ -1,0 +1,169 @@
+use super::workspace_guard::WorkspaceGuard;
+use super::{TransientWorkspaceId, Workspace};
+use crate::api::WindowsApi;
+use crate::common::{Monitor, Monitors, PersistentWorkspaceId, Window, WindowHandle};
+use crate::files::{FileManager, FileType, WorkspacesFile};
+use std::collections::{HashMap, HashSet};
+
+const WORKSPACES_FILE_NAME: &str = "workspaces.toml";
+const WORKSPACE_FILE_PREFIX: &str = "# This file is automatically generated and updated by Randolf.\n\
+  # Please do NOT modify this file manually.\n\
+  # It is used to store references (window handles) to hidden windows.\n\
+  # The file is used by Randolf to restore hidden windows after a crash.\n\n";
+
+pub struct WorkspaceManager<T: WindowsApi> {
+  pub(crate) workspaces: HashMap<PersistentWorkspaceId, Workspace>,
+  pub(crate) windows_api: T,
+  window_margin: i32,
+  additional_workspace_count: i32,
+  pub(crate) file_manager: FileManager<WorkspacesFile>,
+  pub(crate) workspace_file: WorkspacesFile,
+}
+
+impl<T: WindowsApi + Clone> WorkspaceManager<T> {
+  pub fn new(additional_workspace_count: i32, window_margin: i32, api: T) -> Self {
+    let mut file_manager = FileManager::new(WORKSPACES_FILE_NAME, FileType::Data);
+    file_manager.set_content_prefix(WORKSPACE_FILE_PREFIX);
+    let mut workspace_manager = Self {
+      workspaces: HashMap::new(),
+      windows_api: api,
+      window_margin,
+      additional_workspace_count,
+      file_manager,
+      workspace_file: WorkspacesFile::new(),
+    };
+    workspace_manager.initialise_workspaces();
+    workspace_manager.restore_hidden_windows_from_file();
+
+    workspace_manager
+  }
+
+  fn initialise_workspaces(&mut self) {
+    let mut workspaces = HashMap::new();
+    let all_monitors = self.windows_api.get_all_monitors();
+    for monitor in all_monitors.get_all() {
+      if monitor.is_primary {
+        for layer in 1..=self.additional_workspace_count + 1 {
+          let id = PersistentWorkspaceId::new(monitor.id, layer as usize, true);
+          let workspace = if layer == 1 {
+            Workspace::new_active(id, monitor, self.window_margin)
+          } else {
+            Workspace::new_inactive(id, monitor, self.window_margin)
+          };
+          workspaces.insert(id, workspace);
+        }
+      } else {
+        let id = PersistentWorkspaceId::new(monitor.id, 1, false);
+        workspaces.insert(id, Workspace::new_active(id, monitor, self.window_margin));
+      }
+    }
+    self.workspaces = workspaces;
+  }
+
+  pub(crate) fn create_workspace_id_map(&self, monitors: Monitors) -> HashMap<PersistentWorkspaceId, TransientWorkspaceId> {
+    let mut workspace_id_map = HashMap::new();
+    for workspace in self.workspaces.values() {
+      if let Some(monitor) = monitors.get_by_id(&workspace.id.monitor_id) {
+        workspace_id_map.insert(workspace.id, TransientWorkspaceId::from(workspace.id, monitor.handle));
+      }
+    }
+
+    workspace_id_map
+  }
+
+  /// This function will read the workspaces file and attempt to restore any window handles that are present in the
+  /// file. This allows Randolf to recover after any form of non-graceful exit. The workspaces file is continuously
+  /// updated by Randolf.
+  fn restore_hidden_windows_from_file(&mut self) {
+    let (workspace_file, _) = self
+      .file_manager
+      .load_or_create()
+      .expect("Failed to load or create workspaces file");
+    self.workspace_file = workspace_file;
+    info!("Loaded workspaces file: {}", self.workspace_file);
+    let windows = self
+      .windows_api
+      .get_all_windows()
+      .into_iter()
+      .filter(|window| {
+        self
+          .workspace_file
+          .workspaces
+          .values()
+          .any(|handles| handles.contains(&window.handle))
+      })
+      .collect::<HashSet<Window>>();
+    info!("Restoring [{}] hidden windows: {:?}", windows.len(), windows);
+    for window in windows {
+      self.windows_api.do_unhide_window(window.handle);
+    }
+    self.workspace_file.clear(&self.file_manager);
+  }
+
+  pub fn get_ordered_permanent_workspace_ids(&mut self) -> Vec<PersistentWorkspaceId> {
+    let guard = WorkspaceGuard::new(self);
+    guard.get_ordered_workspace_ids()
+  }
+
+  pub fn switch_workspace(&mut self, target_workspace_id: PersistentWorkspaceId) {
+    self.switch_workspace_with_additional_windows(target_workspace_id, &[]);
+  }
+
+  /// Switches workspace while capturing supplied off-screen members.
+  pub fn switch_workspace_with_additional_windows(
+    &mut self,
+    target_workspace_id: PersistentWorkspaceId,
+    additional_windows: &[WindowHandle],
+  ) {
+    let mut guard = WorkspaceGuard::new(self);
+    guard.switch_workspace_with_additional_windows(target_workspace_id, additional_windows);
+  }
+
+  /// Returns the active workspace containing a window's monitor.
+  pub fn active_workspace_for_window(&self, handle: WindowHandle) -> Option<PersistentWorkspaceId> {
+    if self.workspaces.is_empty() {
+      return None;
+    }
+    let monitor_handle = self.windows_api.get_monitor_handle_for_window_handle(handle);
+    let monitor_id = self.windows_api.get_monitor_id_for_handle(monitor_handle)?;
+    self
+      .workspaces
+      .iter()
+      .find_map(|(id, workspace)| (workspace.is_active() && id.monitor_id == monitor_id).then_some(*id))
+  }
+
+  /// Returns all active workspace IDs.
+  pub fn active_workspace_ids(&self) -> Vec<PersistentWorkspaceId> {
+    let mut ids = self
+      .workspaces
+      .iter()
+      .filter_map(|(id, workspace)| workspace.is_active().then_some(*id))
+      .collect::<Vec<_>>();
+    ids.sort();
+    ids
+  }
+
+  /// Returns a workspace's current monitor.
+  pub fn monitor_for_workspace(&self, id: PersistentWorkspaceId) -> Option<Monitor> {
+    self.workspaces.get(&id).map(|workspace| workspace.monitor.clone())
+  }
+
+  /// Returns whether a workspace is active.
+  pub fn is_workspace_active(&self, id: PersistentWorkspaceId) -> bool {
+    self.workspaces.get(&id).is_some_and(Workspace::is_active)
+  }
+
+  pub fn move_window_to_workspace(&mut self, target_workspace_id: PersistentWorkspaceId) {
+    let mut guard = WorkspaceGuard::new(self);
+    guard.move_window_to_workspace(target_workspace_id);
+  }
+
+  pub fn restore_all_managed_windows(&mut self) {
+    let mut guard = WorkspaceGuard::new(self);
+    guard.restore_all_managed_windows();
+  }
+}
+
+#[cfg(test)]
+#[path = "tests/workspace_manager_tests.rs"]
+pub mod tests;
